Update README.md
Thoughts on having a refactor situation here? Essentially, working on reducing the number of event listeners (specifically for the buttons) and then creating a handler function w/ conditional logic to determine which element to add? 

Happy to change this to Iteration 3 or scrap altogether, just figured I'd get your two cents!